fast_spliter
