--- lesson6/zip_newdata.csv
@@ -1,588 +1,0 @@
-70116
-48116
-08014
-99501
-45011
-44805
-60632
-95111
-57105
-21224
-19443
-37275
-11953
-90034
-44023
-78045
-85013
-69734
-37110
-53207
-48180
-61109
-19014
-95111
-75062
-12204
-08846
-54481
-66218
-21601
-90991
-10011
-77301
-74874
-43215
-88011
-07660
-08812
-10025
-70002
-10011
-93012
-78204
-67410
-97754
-66204
-99708
-33196
-99712
-55343
-38938
-02128
-90006
-53711
-19132
-10003
-37388
-98839
-29201
-19087
-08822
-11590
-14288
-19046
-91405
-65895
-02909
-14302
-19006
-02904
-53075
-08831
-78731
-80126
-53214
-10009
-99515
-16502
-21061
-83707
-94104
-27514
-94070
-20932
-94520
-43140
-74989
-14895
-21215
-47565
-07105
-60647
-07104
-88101
-10309
-11279
-32254
-94545
-44122
-76040
-47154
-90247
-60201
-44302
-19106
-94010
-83649
-91776
-76708
-99501
-99385
-95110
-68556
-94577
-46202
-82901
-22102
-70112
-80303
-94577
-96817
-55337
-27263
-11563
-97224
-53511
-01602
-32860
-33133
-16502
-75149
-37388
-07501
-33030
-55892
-21117
-91362
-96819
-31701
-02210
-07601
-94403
-52777
-91761
-72868
-94606
-12180
-18411
-07050
-90131
-08867
-32922
-10468
-96782
-80231
-07866
-08201
-30553
-07062
-02346
-60638
-97202
-75227
-60604
-45840
-92501
-76707
-37211
-53216
-60480
-21655
-75061
-90212
-54169
-11758
-08401
-45203
-12270
-97401
-27577
-10002
-85017
-55262
-46240
-13214
-08002
-11530
-72202
-55101
-63381
-19134
-70360
-11779
-95827
-83704
-78664
-19123
-96720
-89502
-60090
-07869
-29301
-21074
-10553
-10011
-94710
-01742
-94104
-07652
-94561
-60623
-07009
-43551
-47939
-90248
-34448
-95054
-18201
-07304
-94583
-08807
-11716
-91362
-08876
-97005
-97302
-08077
-30135
-32216
-37855
-04401
-76060
-42754
-14228
-33054
-10038
-19103
-32536
-94107
-95132
-94080
-91325
-19103
-70506
-48126
-78754
-99586
-75207
-98070
-19320
-19143
-90248
-94928
-32803
-92025
-01581
-77132
-77301
-53226
-91731
-10701
-75684
-75227
-39530
-33134
-38062
-10048
-07446
-48103
-11729
-44707
-06511
-22030
-01887
-43613
-98409
-95661
-66697
-82501
-96263
-04864
-18954
-33614
-92020
-77840
-60035
-08401
-08807
-25657
-11226
-63104
-95207
-07424
-87393
-46220
-98133
-27215
-67601
-32822
-55401
-29201
-95407
-60067
-26659
-92626
-80212
-06515
-98021
-44136
-66552
-03865
-32937
-25346
-10016
-92647
-48075
-46222
-76301
-78112
-06473
-96950
-21001
-70130
-23219
-18966
-55379
-42744
-33136
-53209
-61304
-08873
-58102
-89701
-85254
-37211
-21801
-10004
-79109
-90016
-91942
-30328
-57254
-01602
-49120
-07004
-70508
-17517
-63088
-70099
-76013
-33619
-30328
-13252
-92234
-48933
-96541
-85034
-04401
-47857
-20735
-17404
-75150
-59701
-07039
-48329
-44103
-79601
-76107
-11230
-02903
-17601
-02919
-33014
-07304
-85092
-78204
-11219
-91768
-44103
-87163
-10013
-23219
-30753
-70518
-07110
-13850
-07055
-83843
-94104
-02760
-48307
-54301
-21230
-80937
-43607
-20785
-02745
-22102
-21076
-78753
-85381
-44820
-98032
-45251
-47404
-60004
-61047
-11223
-32801
-90210
-70295
-33511
-92899
-79925
-92020
-20710
-72119
-92801
-63102
-78664
-27409
-11727
-07071
-38112
-30342
-94290
-94044
-55337
-10759
-85260
-11101
-30340
-55057
-49440
-06854
-46802
-39307
-48310
-80919
-10013
-94545
-60624
-63728
-18509
-95965
-12771
-44142
-07003
-99501
-19102
-45217
-35433
-90248
-50315
-19406
-27601
-99352
-10803
-21234
-11801
-01603
-74105
-49442
-70506
-43402
-87895
-91942
-46314
-06610
-07032
-44203
-07631
-49546
-39208
-38118
-80110
-38112
-91106
-60657
-11103
-99352
-21202
-32820
-07104
-90706
-08755
-07834
-76598
-80216
-23608
-86350
-85012
-33155
-85013
-27846
-32759
-10276
-94105
-92110
-07306
-92410
-02138
-92025
-90248
-17110
-85204
-33461
-90177
-37211
-41005
-48160
-80312
-77301
-07011
-07087
-33142
-38575
-11360
-92705
-26849
-60008
-19440
-07981
-16501
-51120
-84115
-28301
-90021
-48823
-13501
-55713
-46601
-75698
-32806
-07029
-08611
-53717
-10536
-88827
-27401
-21217
-62260
-08831
-78028
-60007
-07728
-21575
-68124
-22003
-75075
-10016
-11791
-34429
-46040
-64504
-63517
-39211
-79602
-95070
-43496
-08003
-94545
-50126
-32750
-38773
-82001
-16452
-11530
-53715
-20113
-07032
-18518
-92126
-92276
-08402
-02906
-67410
-95678
-32114
-20001
-94945
-83709
-46514
-89992
-69301
-61556
-98104
-32804
